First version of minibatches. See notes.
This version of mini-batch SGD changes the previous parallelism level,
so now we are only using as many threads as we set our minibatch size to.

This is obviously pretty bad in terms of utilization for small batch sizes,
so ideally we might want to run many different batches
in parallel and average over those, but it's probably too much work.

Next up: Limit size of gradient vectors to ensure good space usage.